breaks out Avatar and Subcomponents
